Split JVMCloudService into DefaultMinecraftCloudService
